Create command-line functions for maintenance tasks
Tried to build in processes to schedule cron jobs to run maintenance tasks on a regular basis but had trouble either incorporating the logic or tasks would run multiple times depending on the number of workers. Created as command line functions so can be scheduled via regular crontab jobs instead.